+++ src/ScottPlot4/ScottPlot/Renderable/Axis.cs
@@ -27,12 +27,12 @@ namespace ScottPlot.Renderable
         /// <summary>
         /// Axis dimensions and methods for pixel/unit conversions
         /// </summary>
-        public readonly AxisDimensions Dims = new AxisDimensions();
+        public readonly AxisDimensions Dims = new();
 
         /// <summary>
         /// Plottables with this axis index will use pixel/unit conversions from this axis
         /// </summary>
-        public int AxisIndex = 0;
+        public int AxisIndex { get; } = 0;
 
         public bool IsVisible { get; set; } = true;
 
@@ -40,19 +40,22 @@ namespace ScottPlot.Renderable
         public Edge Edge
         {
             get => _Edge;
-            set
+            private set
             {
                 _Edge = value;
                 AxisLine.Edge = value;
                 AxisLabel.Edge = value;
                 AxisTicks.Edge = value;
-                bool isVertical = (value == Edge.Left || value == Edge.Right);
-                AxisTicks.TickCollection.Orientation = isVertical ? AxisOrientation.Vertical : AxisOrientation.Horizontal;
-                Dims.IsInverted = isVertical;
+                AxisTicks.TickCollection.Orientation = value.IsVertical()
+                    ? AxisOrientation.Vertical
+                    : AxisOrientation.Horizontal;
+                Dims.IsInverted = value.IsVertical();
             }
         }
-        public bool IsHorizontal => Edge == Edge.Top || Edge == Edge.Bottom;
-        public bool IsVertical => Edge == Edge.Left || Edge == Edge.Right;
+
+        public bool IsHorizontal => Edge.IsHorizontal();
+
+        public bool IsVertical => Edge.IsVertical();
 
         /// <summary>
         /// Customization options for the text label displayed on an axis
@@ -68,6 +71,12 @@ namespace ScottPlot.Renderable
         /// Customization options for the line between an axis and the data area
         /// </summary>
         public readonly AxisLine AxisLine = new();
+
+        public Axis(int index, Edge edge)
+        {
+            AxisIndex = index;
+            Edge = edge;
+        }
 
         /// <summary>
         /// Return configuration objects to allow deep customization of axis settings.