+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * WpfPlot: Converted the `UserControl` to a `CustomControl` to facilitate inheritance and theming (#2509, #2526) _Thanks @KroMignon_
 * Lollipop and Cleveland plots: Added `LineWidth` property (#2556) _Thanks @benton-anderson_
 * Pie: Added `SliceLabelPosition` property to allow slice labels to be placed outside the pie (#2515, #2510) _Thanks @nuelle16_
+* Axis: Made `Edge` and `AxisIndex` immutable to prevent accidental modification after construction (#2539, #2538) _Thanks @cxjcqu_
+* Plot: Created `LeftAxis`, `RightAxis`, `BottomAxis`, and `TopAxis` which alias `YAxis`, `YAxis2`, `XAxis`, and `XAxis2` but are more expressive (#2568)
 
 ## ScottPlot 5.0.4-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2023-04-02_